refactor: step4패키지 클래스들 domain 패키지로 이동
