+++ src/test/java/nextstep/ladder/step4/DirectionTest.java
@@ -1,5 +1,6 @@
 package nextstep.ladder.step4;
 
+import nextstep.ladder.domain.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
