+++ src/test/java/nextstep/ladder/step4/PointTest.java
@@ -1,5 +1,6 @@
 package nextstep.ladder.step4;
 
+import nextstep.ladder.domain.Point;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
